+++ iotdb-api/pipe-api/src/main/java/org/apache/iotdb/pipe/api/customizer/configuration/PipeConnectorRuntimeConfiguration.java
@@ -17,9 +17,6 @@
  * under the License.
  */
 
-package org.apache.iotdb.pipe.api.customizer.collector;
+package org.apache.iotdb.pipe.api.customizer.configuration;
 
-import org.apache.iotdb.pipe.api.customizer.PipeRuntimeConfiguration;
-
-// TODO: complete this class
-public interface PipeCollectorRuntimeConfiguration extends PipeRuntimeConfiguration {}
+public interface PipeConnectorRuntimeConfiguration extends PipeRuntimeConfiguration {}
