+++ server/src/main/java/org/apache/iotdb/db/pipe/task/stage/PipeTaskConnectorStage.java
@@ -19,13 +19,12 @@
 
 package org.apache.iotdb.db.pipe.task.stage;
 
-import org.apache.iotdb.common.rpc.thrift.TConsensusGroupId;
 import org.apache.iotdb.db.pipe.config.plugin.env.PipeTaskRuntimeEnvironment;
 import org.apache.iotdb.db.pipe.execution.executor.PipeSubtaskExecutorManager;
 import org.apache.iotdb.db.pipe.task.connection.BoundedBlockingPendingQueue;
 import org.apache.iotdb.db.pipe.task.subtask.PipeConnectorSubtaskManager;
-import org.apache.iotdb.pipe.api.customizer.PipeParameters;
-import org.apache.iotdb.pipe.api.customizer.PipeRuntimeEnvironment;
+import org.apache.iotdb.pipe.api.customizer.configuration.PipeRuntimeEnvironment;
+import org.apache.iotdb.pipe.api.customizer.parameter.PipeParameters;
 import org.apache.iotdb.pipe.api.event.Event;
 import org.apache.iotdb.pipe.api.exception.PipeException;
 
@@ -36,14 +35,11 @@ public class PipeTaskConnectorStage extends PipeTaskStage {
   protected String connectorSubtaskId;
 
   public PipeTaskConnectorStage(
-      String pipeName,
-      long creationTime,
-      TConsensusGroupId dataRegionId,
-      PipeParameters pipeConnectorParameters) {
+      String pipeName, long creationTime, PipeParameters pipeConnectorParameters) {
     this.pipeConnectorParameters = pipeConnectorParameters;
 
     final PipeRuntimeEnvironment pipeRuntimeEnvironment =
-        new PipeTaskRuntimeEnvironment(pipeName, creationTime, dataRegionId.getId());
+        new PipeTaskRuntimeEnvironment(pipeName, creationTime);
     connectorSubtaskId =
         PipeConnectorSubtaskManager.instance()
             .register(
